Multiliingual document: titles and labels
Add ability to use content control for titles
(title, shortTitle, subTitle, secTitle) and labels
(acknowledgements, abstract, preface, contents, keywordsShownHere,
endnotes, appendix, glossary, references, glossaryTerms, abbreviations,
selectedBibliography, index).  This enables one to have a
multi-lingual document and be able to choose which language is output
via content control.

Remove ability to have more than one contents (now handled by the above).

In the web page output with a style sheet:
a) Fix bug: keywords could show twice.
b) Fix bug: indexed item in appendix (but not in an embedded section1
or section2) failed to show the Appendix letter in the index output

Fix bug: an empty ul or ol element could either shift remaining text
to the right or insert a blank space.

Fix bug: in the FO PDF and in all style sheet outputs, any shortTitle
in a part was coming out as content in the part.

Fix bug: when using endnotes, an endnote in a part could be listed
with the endnotes in whatever came before the part.