Add ability to have contents in the back matter instead of in the front matter. This is controlled by a publisher style sheet.
(XLIN-14, XLIN-62)

Also, allow the presence of a contentsLayout element in a style sheet
to be controlled by content control.  This helps to let one have a single
document where the content is in different languages and one can control
whether the table of contents appears in the front matter (e.g., English)
or in the back matter (e.g., French) via the content control for that
language.