Fix two bugs for XeLaTeX PDF:
When a blockquote was immediately followed by an ordered list (ol) or
an unordered list (ul), there was too much blank space between the two.

In a style sheet, when a section1 element was the only section or chapter
element in the document and the useperiodafternumber attribute of the
sectionTitleLayout element was set to 'yes', there would be a period
before the section title.