Fix broken link in the JIPA Illustration submission template.
For the default way of producing PDF (via XeLaTeX):
Fix bug: when using a publisher style sheet, an img element with a
XeLaTeXSpecial attribute of vertical-adjustment could cause the PDF to fail to be produced.

Fix bug: when a publisher style sheet had a background color set
for a title, the PDF could fail to be produced.