XLIN-95 and XLIN-98
Add a way to control author surname and given names in collections and
proceedings via collCitation and procCitation.

Add ability to control abbreviation/definition display in footnotes.

Attempt to have the automatic spelling checker use the xml:lang attribute
of the lingPaper element as a first attempt at determining what language
to use for spelling checking.